refactor: extract camera_ready_campaign config dataclasses into config module (#3405)
First slice of the #3385 decomposition of the ~4967-line camera_ready_campaign.py
(largest module in robot_sf/). Extracts the lowest-risk, highest-clarity cluster.

- Move the 6 frozen config dataclasses (AmvProfileConfig, SeedPolicy,
  ScenarioCandidateSelection, PlannerSpec, SnqiContractConfig, CampaignConfig) plus
  the _AMV_DIMENSIONS and DEFAULT_SEED_SETS_PATH constants their defaults reference
  into robot_sf/benchmark/camera_ready_campaign_config.py.
- Re-export all of them from camera_ready_campaign so existing import paths keep
  working (release_protocol.py, orca_preflight.py, tests).
- The two cross-types (SyntheticActuationProfile, LatencyStressProfile) are only used
  in CampaignConfig annotations, so the config module imports them under TYPE_CHECKING
  from their canonical modules; the config module imports nothing from
  camera_ready_campaign, keeping the graph acyclic.
- Behavior-preserving verbatim move: test_camera_ready_campaign, test_run_benchmark_release,
  test_orca_preflight, test_analyze/compare_camera_ready_campaigns, and the research-v1
  matrix tests all pass.

Out of scope (later children of #3385): load_campaign_config/_validate_campaign_config,
the _write_* artifact-IO, _build_* summary, route-clearance, and reporting clusters.
